Banquet E.G.O. weapon rework and reflavour (#110)
* Banquet E.G.O. weapon reflavour

Used to be a vaguely nosferatu themed cannon that shot fireballs, now is a staff that shoots blood spikes. New obj and inhand sprites, new projectile sprite, new fire noise, aesthetic bloodsplatters when fired and on impact.

* Bat summoning ability

This MAY be a little deranged? Can now use bloodfeast to alt click the weapon, spawns a friendly bat. Hitting an enemy with the projectile will force any spawned bats to target that enemy

* Slight rebalancing

This should be fine :clueless:

* Adds missing null check

* Apply suggestions from code review

Co-authored-by: Gboster-0 <82319946+Gboster-0@users.noreply.github.com>

* Adds a null check to bat's Destroy() and minimizes operations in Banquet's Destroy()

As advised by Gboster

* Replaces bound_staff var on bats with use of signals on the weapon

* Nulls bat master var on destroy, also unregisters the signal in the cleanup proc

* Marks DestructionCleanup as a signal handler

I forgot I had to do this

* Resolves icon conflicts

It was probably Encompassing

---------

Co-authored-by: Gboster-0 <82319946+Gboster-0@users.noreply.github.com>